Added a gitignore, removed backspace changes and added C++ snippets
 * Added functionality for C++0x, such as auto assign, templates, range
   based for loops, lambdas (empty, by value, and by reference), new
   method and function syntax ( auto name(params) -> return_type), raw
   strings, the new use of "using", and delete.
 * For C, changed the custom for loop to use `forc` rather than `forr`,
   made the default function declaration and definition parameters to
   have "void" placed inside of them (as per the standard), and added a
   "sizeof" snippet with sz.

Should be adding more for the C++ functionality soon, such as the
include guard ending in HPP, as well as iterator based loops, a "ref"
and "mov" snippet for add parameters by reference or the new move (&&)
operator, as well as operator overload snippets.

Will obviously be adding more as ideas come to me.